+++ .eslintrc.json
@@ -2,8 +2,8 @@
   "extends": ["next/core-web-vitals", "next/typescript"],
   "rules": {
     "@typescript-eslint/no-unused-vars": "off",
-     "@typescript-eslint/no-explicit-any": "off",
-      "react-hooks/rules-of-hooks": "off", 
+    "@typescript-eslint/no-explicit-any": "off",
+    "react-hooks/rules-of-hooks": "off",
     "react-hooks/exhaustive-deps": "off"
   }
 }
